migrate file selector controller tests to integration
